Rename theme name from MangaSeeLife to NepNep
