+++ multisrc/overrides/nepnep/mangalife/res/mipmap-hdpi/ic_launcher.png
