+++ multisrc/overrides/nepnep/mangalife/res/mipmap-mdpi/ic_launcher.png
