+++ multisrc/overrides/nepnep/mangasee/res/mipmap-hdpi/ic_launcher.png
