+++ multisrc/overrides/nepnep/mangasee/res/mipmap-mdpi/ic_launcher.png
