+++ multisrc/overrides/nepnep/mangasee/res/mipmap-xxxhdpi/ic_launcher.png
